BitCount datatype changed!

+++ response.go
@@ -2,5 +2,5 @@ package idgenmodel
 
 type IDGenResp struct {
 	ID       []byte `json:"id"`
-	BitCount uint   `json:"bit-count"`
+	BitCount int64  `json:"bit-count"`
 }
